+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/KafkaVersionTestUtils.java
@@ -21,22 +21,22 @@ public class KafkaVersionTestUtils {
     public static final String KAFKA_MIRROR_MAKER_IMAGE_STR = "strimzi/kafka-mirror-maker:latest-kafka-";
     public static final String KAFKA_MIRROR_MAKER_2_IMAGE_STR = "strimzi/kafka-connect:latest-kafka-";
 
-    public static final String LATEST_KAFKA_VERSION = "2.4.1";
-    public static final String LATEST_FORMAT_VERSION = "2.4";
-    public static final String LATEST_PROTOCOL_VERSION = "2.4";
+    public static final String LATEST_KAFKA_VERSION = "2.5.0";
+    public static final String LATEST_FORMAT_VERSION = "2.5";
+    public static final String LATEST_PROTOCOL_VERSION = "2.5";
     public static final String LATEST_ZOOKEEPER_VERSION = "3.5.7";
     public static final String LATEST_CHECKSUM = "ABCD1234";
-    public static final String LATEST_THIRD_PARTY_VERSION = "2.4.x";
+    public static final String LATEST_THIRD_PARTY_VERSION = "2.5.x";
     public static final String LATEST_KAFKA_IMAGE = KAFKA_IMAGE_STR + LATEST_KAFKA_VERSION;
     public static final String LATEST_KAFKA_CONNECT_IMAGE = KAFKA_CONNECT_IMAGE_STR + LATEST_KAFKA_VERSION;
     public static final String LATEST_KAFKA_CONNECT_S2I_IMAGE = KAFKA_CONNECT_S2I_IMAGE_STR + LATEST_KAFKA_VERSION;
     public static final String LATEST_KAFKA_MIRROR_MAKER_IMAGE = KAFKA_MIRROR_MAKER_IMAGE_STR + LATEST_KAFKA_VERSION;
     public static final String LATEST_KAFKA_MIRROR_MAKER_2_IMAGE = KAFKA_MIRROR_MAKER_2_IMAGE_STR + LATEST_KAFKA_VERSION;
 
-    public static final String LATEST_MINOR_KAFKA_VERSION = "2.4.0";
-    public static final String LATEST_MINOR_FORMAT_VERSION = "2.4";
-    public static final String LATEST_MINOR_PROTOCOL_VERSION = "2.4";
-    public static final String LATEST_MINOR_ZOOKEEPER_VERSION = "3.5.6";
+    public static final String LATEST_MINOR_KAFKA_VERSION = "2.5.0";
+    public static final String LATEST_MINOR_FORMAT_VERSION = "2.5";
+    public static final String LATEST_MINOR_PROTOCOL_VERSION = "2.5";
+    public static final String LATEST_MINOR_ZOOKEEPER_VERSION = "3.5.7";
     public static final String LATEST_MINOR_CHECKSUM = "ABCD1234";
     public static final String LATEST_MINOR_THIRD_PARTY_VERSION = "2.4.x";
     public static final String LATEST_MINOR_KAFKA_IMAGE = KAFKA_IMAGE_STR + LATEST_MINOR_KAFKA_VERSION;
@@ -45,24 +45,24 @@ public class KafkaVersionTestUtils {
     public static final String LATEST_MINOR_KAFKA_MIRROR_MAKER_IMAGE = KAFKA_MIRROR_MAKER_IMAGE_STR + LATEST_MINOR_KAFKA_VERSION;
     public static final String LATEST_MINOR_KAFKA_MIRROR_MAKER_2_IMAGE = KAFKA_MIRROR_MAKER_2_IMAGE_STR + LATEST_MINOR_KAFKA_VERSION;
 
-    public static final String PREVIOUS_KAFKA_VERSION = "2.3.1";
-    public static final String PREVIOUS_FORMAT_VERSION = "2.3";
-    public static final String PREVIOUS_PROTOCOL_VERSION = "2.3";
-    public static final String PREVIOUS_ZOOKEEPER_VERSION = "3.4.14";
+    public static final String PREVIOUS_KAFKA_VERSION = "2.4.1";
+    public static final String PREVIOUS_FORMAT_VERSION = "2.4";
+    public static final String PREVIOUS_PROTOCOL_VERSION = "2.4";
+    public static final String PREVIOUS_ZOOKEEPER_VERSION = "3.5.6";
     public static final String PREVIOUS_CHECKSUM = "ABCD1234";
-    public static final String PREVIOUS_THIRD_PARTY_VERSION = "2.3.x";
+    public static final String PREVIOUS_THIRD_PARTY_VERSION = "2.4.x";
     public static final String PREVIOUS_KAFKA_IMAGE = KAFKA_IMAGE_STR + PREVIOUS_KAFKA_VERSION;
     public static final String PREVIOUS_KAFKA_CONNECT_IMAGE = KAFKA_CONNECT_IMAGE_STR + PREVIOUS_KAFKA_VERSION;
     public static final String PREVIOUS_KAFKA_CONNECT_S2I_IMAGE = KAFKA_CONNECT_S2I_IMAGE_STR + PREVIOUS_KAFKA_VERSION;
     public static final String PREVIOUS_KAFKA_MIRROR_MAKER_IMAGE = KAFKA_MIRROR_MAKER_IMAGE_STR + PREVIOUS_KAFKA_VERSION;
     public static final String PREVIOUS_KAFKA_MIRROR_MAKER_2_IMAGE = KAFKA_MIRROR_MAKER_2_IMAGE_STR + PREVIOUS_KAFKA_VERSION;
 
-    public static final String PREVIOUS_MINOR_KAFKA_VERSION = "2.3.0";
-    public static final String PREVIOUS_MINOR_FORMAT_VERSION = "2.3";
-    public static final String PREVIOUS_MINOR_PROTOCOL_VERSION = "2.3";
-    public static final String PREVIOUS_MINOR_ZOOKEEPER_VERSION = "3.4.14";
+    public static final String PREVIOUS_MINOR_KAFKA_VERSION = "2.4.0";
+    public static final String PREVIOUS_MINOR_FORMAT_VERSION = "2.4";
+    public static final String PREVIOUS_MINOR_PROTOCOL_VERSION = "2.4";
+    public static final String PREVIOUS_MINOR_ZOOKEEPER_VERSION = "3.5.6";
     public static final String PREVIOUS_MINOR_CHECKSUM = "ABCD1234";
-    public static final String PREVIOUS_MINOR_THIRD_PARTY_VERSION = "2.3.x";
+    public static final String PREVIOUS_MINOR_THIRD_PARTY_VERSION = "2.4.x";
     public static final String PREVIOUS_MINOR_KAFKA_IMAGE = KAFKA_IMAGE_STR + PREVIOUS_MINOR_KAFKA_VERSION;
     public static final String PREVIOUS_MINOR_KAFKA_CONNECT_IMAGE = KAFKA_CONNECT_IMAGE_STR + PREVIOUS_MINOR_KAFKA_VERSION;
     public static final String PREVIOUS_MINOR_KAFKA_CONNECT_S2I_IMAGE = KAFKA_CONNECT_S2I_IMAGE_STR + PREVIOUS_MINOR_KAFKA_VERSION;
@@ -141,7 +141,6 @@ public class KafkaVersionTestUtils {
     public static String getKafkaVersionYaml() {
         return getPreviousMinorVersionYaml(false) +
                 getPreviousVersionYaml(false) +
-                getLatestMinorVersionYaml(false) +
                 getLatestVersionYaml(true);
     }